Fixed some bugs in the extension of the mesh. Solved the issue of extension and remeshing within one time step.

+++ src/controller.py
@@ -328,7 +328,7 @@ class Controller:
                         print("Remeshing by extending in the vertical direction the domain...")
 
                         # AM:
-                        elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor * 2 - 1)
+                        elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor * 2 - 1))
                         if elems_add % 2 != 0:
                             elems_add = elems_add + 1
 
@@ -361,7 +361,7 @@ class Controller:
                         print("Remeshing by extending in the horizontal direction the domain...")
 
                         # AM:
-                        elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor * 2 - 1)
+                        elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor * 2 - 1))
                         if elems_add % 2 != 0:
                             elems_add = elems_add + 1
 
@@ -396,7 +396,7 @@ class Controller:
                             print("Remeshing by extending towards negative y, positive y hit also so extending by half "
                                   "the factor in vertical direction...")
                             # AM:
-                            elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -408,7 +408,7 @@ class Controller:
                             print("Remeshing by extending towards negative y...")
 
                             # AM:
-                            elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -443,7 +443,7 @@ class Controller:
                             print("Remeshing by extending towards positive y, negative y hit also so extending by half "
                                   "the factor in vertical direction...")
                             # AM:
-                            elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -455,7 +455,7 @@ class Controller:
                             print("Remeshing by extending towards positive y...")
 
                             # AM:
-                            elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -490,7 +490,7 @@ class Controller:
                             print("Remeshing by extending towards negative x, positive x hit also so extending by half "
                                   "the factor in vertical direction...")
                             # AM:
-                            elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -502,7 +502,7 @@ class Controller:
                             print("Remeshing by extending towards negative x...")
 
                             # AM:
-                            elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -537,7 +537,7 @@ class Controller:
                             print("Remeshing by extending towards positive x, negative x hit also so extending by half "
                                   "the factor in vertical direction...")
                             # AM:
-                            elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -549,7 +549,7 @@ class Controller:
                             print("Remeshing by extending towards positive x...")
 
                             # AM:
-                            elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
